Refactor code

+++ src/Component/NoteList.tsx
@@ -1,35 +1,35 @@
 import React, { useEffect, useState } from "react";
-import { Note } from "../Type/NoteListPropsType";
+import { Note } from "../Type/RelatedNoteType";
 import apiService from '../ApiCaller/ApiCaller';
-import { NoteListPropsType } from "../Type/NoteListPropsType";
-
+import { NoteListPropsType } from "../Type/RelatedNoteType";
+import { Table, TableHead, TableBody, TableRow, TableCell, Button, Box,TextField } from '@mui/material';
 
 function NoteList(props: NoteListPropsType) {
 
 
 
   const handleContentChange = (id: number, newContent: string) => {
-    const updatedNotes = props.notes.map((note) => {
+    const updatedNotes = props.data.value.map((note) => {
       if (note.id === id) {
         return { ...note, content: newContent };
       }
       return note;
     });
-    props.setNotes(updatedNotes);
+    props.data.setValue(updatedNotes);
   };
 
 
   const handleSave = (noteId: number) => {
 
 
-    const noteToUpdate = props.notes.find((note) => note.id === noteId);
+    const noteToUpdate = props.data.value.find((note) => note.id === noteId);
 
-    if (noteToUpdate) {
+    if (noteToUpdate && typeof props.repoId === "number") {
 
       const updateContent = noteToUpdate.content;
 
       apiService.updateNote(props.repoId, noteId, updateContent)
-        .then(() => {props.setUpdateNoteStatus(!props.updateNoteStatus)})
+        .then(() => {props.updatingNoteFlag.setValue(!props.updatingNoteFlag.value)})
         .catch((error)=>{console.error("Error:", error)});
     }  
     else {
@@ -42,12 +42,12 @@ function NoteList(props: NoteListPropsType) {
 
   const handleDelete = (noteId: number) => {
 
-    const noteToDelete = props.notes.find((note) => note.id === noteId);
+    const noteToDelete = props.data.value.find((note) => note.id === noteId);
     
-    if (noteToDelete) {
+    if (noteToDelete && typeof props.repoId === "number") {
 
       apiService.deleteNote(props.repoId, noteId)
-        .then(() => {props.setUpdateNoteStatus(!props.updateNoteStatus)})
+        .then(() => {props.updatingNoteFlag.setValue(!props.updatingNoteFlag.value)})
         .catch((error)=>{console.error("Error:", error)});
 
     } 
@@ -59,34 +59,35 @@ function NoteList(props: NoteListPropsType) {
 
     
     return (
-
-        <div>
-        <table>
-          <thead>
-            <tr>
-                <th>Id</th>
-                <th>Content</th>
-                <th>Action</th>
-            </tr>
-          </thead>
-          <tbody>
-            {props.notes.map((item) => (
-              <tr key={item.id}>
-                <td>{item.id}</td>
-                <td><input
+   <Box>
+      <Table>
+        <TableHead>
+          <TableRow>
+            <TableCell align="center">Id</TableCell>
+            <TableCell align="center">Note</TableCell>
+            <TableCell align="center">Action</TableCell>
+          </TableRow>
+        </TableHead>
+        <TableBody>
+          {props.data.value.map((item, keyId) => (
+            <TableRow key={keyId}>
+              <TableCell>{keyId + 1}</TableCell>
+              <TableCell>
+                <TextField
                   type="text"
                   value={item.content}
                   onChange={(e) => handleContentChange(item.id, e.target.value)}
-                /></td>
-                <td>
-                  <button onClick={() => handleDelete(item.id)}>Delete</button>
-                  <button onClick={() =>handleSave(item.id)}>Save</button>
-              </td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
+                />
+              </TableCell>
+              <TableCell>
+                <Button onClick={() => handleDelete(item.id)}>Delete</Button>
+                <Button onClick={() => handleSave(item.id)}>Save</Button>
+              </TableCell>
+            </TableRow>
+          ))}
+        </TableBody>
+      </Table>
+    </Box>
     )
 
 }
